Inventory WIP
Currently working on the inventory system and have implemented the necessary functions to handle the inventory. Currently, this has not been thoroughly tested and must undergo lots of testing, so this may not work as expected. No visual representation of the inventory is implemented in this push, but that is a primary goal of upcoming commits